% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_table.R
\name{balance_table}
\alias{balance_table}
\title{Makes the columns of the dataset lowercase.}
\usage{
balance_table(df, treatment, variables = names(df))
}
\arguments{
\item{df}{A dataset (data.frame, tibble, etc.).}

\item{treatment}{The treatment variable as a string.}

\item{variables}{The subset of columns to test for balance. Select all columns by default.}
}
\value{
A balance table with columns term, estimate, std.error, statistic, and p.value.
        Term is the variable being tested. Estimate is the distance between the means of treated
        and untreated groups. The p.value is the likelihood that distance is up to chance. In this
        case then null hypothesis is that the treatment and control groups are the same.
}
\description{
Makes the columns of the dataset lowercase.
}
