% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusion_test.R
\name{exclusion_test}
\alias{exclusion_test}
\title{Test for a significant difference between two models.}
\usage{
exclusion_test(unrestricted.model, restricted.model, sig.level = 0.05)
}
\arguments{
\item{unrestricted.model}{A fitted model object.}

\item{restricted.model}{A fitted model object.}

\item{sig.level}{The significance level, default is 0.05.}
}
\value{
TRUE or FALSE
}
\description{
This function takes in two model objects and a significance level.
It returns TRUE if there is a difference between the restricted and
unrestricted model.
}
