% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_names.R
\name{lower_names}
\alias{lower_names}
\title{Makes the columns of the dataset lowercase.}
\usage{
lower_names(df)
}
\arguments{
\item{df}{A dataset (data.frame, tibble, etc.).}
}
\value{
The same dataset, but with lowercase column names.
}
\description{
Makes the columns of the dataset lowercase.
}
\examples{
lower_names(data.frame("A" = c(1,2,3), "B" = seq(1:3)))
}
