% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_rank.R
\name{percentile_rank}
\alias{percentile_rank}
\title{Creates the percentile rank of a vector.}
\usage{
percentile_rank(x)
}
\arguments{
\item{x}{A numeric column.}
}
\value{
The input percentile ranked (vector).
}
\description{
Creates the percentile rank of a vector.
}
\examples{
percentile_rank(seq(1:10))
}
