% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_summary.R
\name{quick_summary}
\alias{quick_summary}
\title{Summarize a dataset.}
\usage{
quick_summary(df)
}
\arguments{
\item{df}{A dataset (data.frame, tibble, etc.).}
}
\value{
Summary information as a data.frame.
}
\description{
This function displays percent missing, class, number of unique values,
and an example value for each column of a dataset.
}
\examples{
quick_summary(cars)
}
