% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{tt}
\alias{tt}
\title{Transpose and return a tibble.}
\usage{
tt(df, new.names = NULL)
}
\arguments{
\item{df}{A dataset (data.frame, tibble, etc.).}

\item{new.names}{A vector of names for the new columns (optional).}
}
\value{
A transposed dataset as a tibble.
}
\description{
Transpose and return a tibble.
}
\examples{
df <- rbind(apply(cars, 2, mean), apply(cars, 2, sd))
tt(df, new.names = c("m", "s"))
}
